CR-660 Split up complete, and all unit tests passing
